added new strange attractor system
