+++ index.html
@@ -35,6 +35,7 @@
           <div class="mb-3 row">
             <select class="form-select" name="equation">
               <option value="lorenz" selected>Lorenz System</option>
+              <option value="rossler" selected>Rossler System</option>
             </select>
           </div>
           <div class="mb-3 row">
